requirements.txt
streamlit
transformers
torch
Pillow
soundfile
sentence-transformers
espnet
espnet_model_zoo
accelerate
git+https://github.com/openai/CLIP.git
flash_attn